Improve pagination navigation with top/bottom controls and scroll-to-top button
- Enhanced Pagination component with compact mobile-friendly design
- Shows current page, adjacent pages, and first/last with ellipsis
- Added pagination at both top and bottom of game results
- Implemented floating "back to top" button that appears after scrolling 400px
- Smooth scroll behavior for better UX
- Fully accessible with proper ARIA labels
- Works seamlessly on mobile, tablet, and desktop

This eliminates the need to scroll back to top after browsing games, significantly improving the user experience.
